backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mssql-django>=1.4
django-filter

